# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "DONE. API",
    "version" : "1.0.0",
    "description" : "Native Apple API. Email/password JWT and native Sign in with Apple. All timestamps UTC; history is server-owned. Durable activity events; APNs transport is not part of this phase."
  },
  "servers" : [ {
    "description" : "Local development; configure PUBLIC_ORIGIN for deployed links.",
    "url" : "http://127.0.0.1:8080"
  } ],
  "components" : {
    "securitySchemes" : {
      "bearerAuth" : {
        "bearerFormat" : "JWT",
        "scheme" : "bearer",
        "type" : "http"
      },
      "writerCapability" : {
        "description" : "43-character base64url private writer key; never a JWT and never included in the public link.",
        "type" : "http",
        "bearerFormat" : "Private write capability",
        "scheme" : "bearer"
      }
    },
    "schemas" : {
      "ActivityView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "anyOf" : [ {
              "format" : "uuid",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "applianceId" : {
            "anyOf" : [ {
              "format" : "uuid",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "sessionId" : {
            "anyOf" : [ {
              "format" : "uuid",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "kind" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "title" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "detail" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "occurredAt" : {
            "anyOf" : [ {
              "format" : "date-time",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "readAt" : {
            "anyOf" : [ {
              "format" : "date-time",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "titleKey" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "message" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/LocalizedMessage"
            }, {
              "type" : "null"
            } ]
          },
          "programSnapshot" : {
            "anyOf" : [ {
              "$ref" : "#/components/schemas/ProgramSnapshot"
            }, {
              "type" : "null"
            } ]
          }
        },
        "additionalProperties" : false,
        "required" : [ "id", "applianceId", "sessionId", "kind", "title", "detail", "occurredAt", "readAt", "titleKey", "message", "programSnapshot" ]
      },
...
      "AppleLogin" : {
        "type" : "object",
        "properties" : {
          "challengeId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "identityToken" : {
            "type" : "string",
            "maxLength" : 16000
          },
          "authorizationCode" : {
            "type" : "string",
            "maxLength" : 4096
          },
          "displayName" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "language" : {
            "anyOf" : [ {
              "type" : "string",
              "pattern" : "en|ro|es|it|fr|de|pl|hi|ja"
            }, {
              "type" : "null"
            } ]
          },
          "clientId" : {
            "anyOf" : [ {
              "type" : "string",
              "maxLength" : 255
            }, {
              "type" : "null"
            } ]
          }
        },
        "additionalProperties" : false,
        "required" : [ "challengeId", "identityToken", "authorizationCode" ]
      },
...
      "ChangePassword" : {
        "type" : "object",
        "properties" : {
          "currentPassword" : {
            "type" : "string",
            "maxLength" : 64
          },
          "newPassword" : {
            "type" : "string",
            "minLength" : 12,
            "maxLength" : 64
          }
        },
        "additionalProperties" : false,
        "required" : [ "currentPassword", "newPassword" ]
      },
...
      "DeleteAccount" : {
        "type" : "object",
        "properties" : {
          "password" : {
            "anyOf" : [ {
              "type" : "string",
              "maxLength" : 64
            }, {
              "type" : "null"
            } ]
          }
        },
        "additionalProperties" : false
      },
...
      "LocalizedMessage" : {
        "type" : "object",
        "properties" : {
          "key" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "arguments" : {
            "anyOf" : [ {
              "items" : {
                "type" : "string"
              },
              "type" : "array"
            }, {
              "type" : "null"
            } ]
          }
        },
        "additionalProperties" : false,
        "required" : [ "key", "arguments" ]
      },
...
      "MutationResult" : {
        "type" : "object",
        "properties" : {
          "revision" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "additionalProperties" : false,
        "required" : [ "revision" ]
      },
      "PageResponse_ActivityView" : {
        "required" : [ "items", "page", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/ActivityView"
            },
            "type" : "array"
          },
          "page" : {
            "format" : "int32",
            "type" : "integer"
          },
          "size" : {
            "format" : "int32",
            "type" : "integer"
          },
          "totalElements" : {
            "format" : "int64",
            "type" : "integer"
          },
          "totalPages" : {
            "format" : "int32",
            "type" : "integer"
          }
        }
      },
...
      "Problem" : {
        "required" : [ "status", "code", "detail" ],
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer"
          },
          "detail" : {
            "type" : "string"
          }
        }
      },
...
      "ProgramSnapshot" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string"
          },
          "minutes" : {
            "format" : "int32",
            "type" : "integer"
          },
          "localizationKey" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "isCustom" : {
            "type" : "boolean"
          }
        },
        "additionalProperties" : false,
        "description" : "Immutable value snapshot. name is literal user content or a built-in label; localizationKey is optional. Measured values are separate from the original snapshot.",
        "required" : [ "name", "minutes", "isCustom" ]
      },
...
      "UpdateProfile" : {
        "type" : "object",
        "properties" : {
          "displayName" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "timezone" : {
            "anyOf" : [ {
              "type" : "string",
              "maxLength" : 64
            }, {
              "type" : "null"
            } ]
          },
          "notificationsEnabled" : {
            "anyOf" : [ {
              "type" : "boolean"
            }, {
              "type" : "null"
            } ]
          },
          "language" : {
            "anyOf" : [ {
              "type" : "string",
              "pattern" : "en|ro|es|it|fr|de|pl|hi|ja"
            }, {
              "type" : "null"
            } ]
          },
          "onboarded" : {
            "anyOf" : [ {
              "type" : "boolean"
            }, {
              "type" : "null"
            } ]
          }
        },
        "additionalProperties" : false
      },
...
      "UserView" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "anyOf" : [ {
              "format" : "uuid",
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "email" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "displayName" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "timezone" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "language" : {
            "anyOf" : [ {
              "type" : "string"
            }, {
              "type" : "null"
            } ]
          },
          "notificationsEnabled" : {
            "type" : "boolean"
          },
          "passwordEnabled" : {
            "type" : "boolean"
          },
          "appleEnabled" : {
            "type" : "boolean"
          },
          "revision" : {
            "format" : "int64",
            "type" : "integer"
          },
          "onboarded" : {
            "type" : "boolean"
          }
        },
        "additionalProperties" : false,
        "required" : [ "id", "email", "displayName", "timezone", "language", "notificationsEnabled", "passwordEnabled", "appleEnabled", "revision", "onboarded" ]
      }
    }
  },
  "paths" : {
...
    "/api/v1/activity" : {
      "get" : {
        "operationId" : "ActivityController_list_3181616031",
        "tags" : [ "Activity" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "parameters" : [ {
          "name" : "page",
          "schema" : {
            "format" : "int32",
            "type" : "integer",
            "default" : 0
          },
          "required" : false,
          "in" : "query"
        }, {
          "name" : "size",
          "schema" : {
            "format" : "int32",
            "type" : "integer",
            "default" : 20
          },
          "required" : false,
          "in" : "query"
        }, {
          "name" : "householdId",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : false,
          "in" : "query"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponse_ActivityView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/activity/read-all" : {
      "post" : {
        "operationId" : "ActivityController_readAll_530642266",
        "tags" : [ "Activity" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "parameters" : [ {
          "name" : "Idempotency-Key",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "header"
        }, {
          "name" : "householdId",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : false,
          "in" : "query"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MutationResult"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/activity/{id}/read" : {
      "post" : {
        "operationId" : "ActivityController_read_1944713166",
        "tags" : [ "Activity" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "parameters" : [ {
          "name" : "id",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "path"
        }, {
          "name" : "Idempotency-Key",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "header"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ActivityView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/me" : {
      "delete" : {
        "operationId" : "AccountController_delete_2200265256",
        "tags" : [ "Account" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DeleteAccount"
              }
            }
          },
          "required" : false
        },
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "204" : {
            "description" : "Success"
          }
        }
      },
      "get" : {
        "operationId" : "AccountController_get_2200265256",
        "tags" : [ "Account" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserView"
                }
              }
            }
          }
        }
      },
      "patch" : {
        "operationId" : "AccountController_update_2200265256",
        "tags" : [ "Account" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateProfile"
              }
            }
          },
          "required" : true
        },
        "parameters" : [ {
          "name" : "Idempotency-Key",
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          },
          "required" : true,
          "in" : "header"
        } ],
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserView"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/me/identities/apple" : {
      "post" : {
        "operationId" : "AccountController_apple_1174696558",
        "tags" : [ "Account" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/AppleLogin"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "200" : {
            "description" : "Success",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserView"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/me/password" : {
      "post" : {
        "operationId" : "AccountController_password_6071970",
        "tags" : [ "Account" ],
        "security" : [ {
          "bearerAuth" : [ ]
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChangePassword"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "default" : {
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/Problem"
                }
              }
            },
            "description" : "Error; code is stable and independent of language."
          },
          "204" : {
            "description" : "Success"
          }
        }
      }
    },
...
  }
}